+++ myokit/_err.py
@@ -419,9 +419,10 @@ class UnusedVariableError(IntegrityError):
         super(UnusedVariableError, self).__init__(msg, tok)
 
 
-class WellMappedError(MyokitError):
+class VariableMappingError(MyokitError):
     """
-    Raised when a variable map is not well defined.
+    Raised when a method needs to map variables from one model onto another,
+    but no valid mapping can be made.
 
     *Extends:* :class:`myokit.MyokitError`.
     """